% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_fiscal_calendar_quarters.R
\name{create_fiscal_yr_Q}
\alias{create_fiscal_yr_Q}
\title{Create Fiscal Year and Fiscal Quarter Columns}
\usage{
create_fiscal_yr_Q(df, date_col)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{date_col}{The unquoted name of the date column.}
}
\value{
A modified data frame with additional fiscal columns.
}
\description{
This function takes a dataframe and a date column to append
fiscal quarter (Q1-Q4) and fiscal year based on a July 1st start date.
}
